% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-manifest.R
\name{analysis_manifest_table}
\alias{analysis_manifest_table}
\title{Summarise Manifest Components}
\usage{
analysis_manifest_table(manifest)
}
\arguments{
\item{manifest}{An analysis manifest.}
}
\value{
A data frame of key provenance components.
}
\description{
Summarise Manifest Components
}
