% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-contract.R
\name{as.data.frame.gp3bayes_pupil_contract}
\alias{as.data.frame.gp3bayes_pupil_contract}
\title{Convert a pupil contract to a transparent table}
\usage{
\method{as.data.frame}{gp3bayes_pupil_contract}(x, ...)
}
\arguments{
\item{x}{A \code{gp3bayes_pupil_contract}.}

\item{...}{Ignored.}
}
\value{
A data frame of contract fields.
}
\description{
Convert a pupil contract to a transparent table
}
