% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-estimands.R
\name{as.data.frame.gp3bayes_pupil_estimand}
\alias{as.data.frame.gp3bayes_pupil_estimand}
\title{Convert a pupil estimand to a data frame}
\usage{
\method{as.data.frame}{gp3bayes_pupil_estimand}(x, ...)
}
\arguments{
\item{x}{A \code{gp3bayes_pupil_estimand}.}

\item{...}{Ignored.}
}
\value{
The estimand table.
}
\description{
Convert a pupil estimand to a data frame
}
