% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{backend_environment_table}
\alias{backend_environment_table}
\title{Backend-Environment Table}
\usage{
backend_environment_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_backend_environment}.}
}
\value{
Backend environment checks.
}
\description{
Backend-Environment Table
}
