% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-support.R
\name{binary_calibration_table}
\alias{binary_calibration_table}
\title{Binary Calibration Table}
\usage{
binary_calibration_table(x, bins = 10L, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{x}{A binary expected-response \code{gp3bayes_prediction}.}

\item{bins}{Number of equal-frequency calibration bins.}

\item{probs}{Posterior interval probabilities.}
}
\value{
A data frame comparing observed event rates with posterior mean
event probabilities by bin.
}
\description{
Binary Calibration Table
}
