% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-distribution-atlas.R
\name{binary_calibration_uncertainty_table}
\alias{binary_calibration_uncertainty_table}
\title{Binary Calibration-Uncertainty Table}
\usage{
binary_calibration_uncertainty_table(x)
}
\arguments{
\item{x}{A binary calibration-uncertainty object.}
}
\value{
Bin-level summaries.
}
\description{
Binary Calibration-Uncertainty Table
}
