% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-diagnostics-advanced.R
\name{binary_group_calibration}
\alias{binary_group_calibration}
\title{Grouped Binary Calibration}
\usage{
binary_group_calibration(x, group)
}
\arguments{
\item{x}{A binary expected-response \code{gp3bayes_prediction} with observed
outcomes.}

\item{group}{Name of a column in \code{x$newdata}.}
}
\value{
A group-level calibration summary.
}
\description{
Grouped Binary Calibration
}
