% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-optional-workflows.R
\name{compute_loo_model_weights}
\alias{compute_loo_model_weights}
\title{Compute LOO Model-Averaging Weights}
\usage{
compute_loo_model_weights(x, method = c("stacking", "pseudobma"), cores = 1L)
}
\arguments{
\item{x}{A \code{gp3bayes_loo_comparison} or named list of PSIS-LOO objects.}

\item{method}{Either stacking or pseudo-BMA.}

\item{cores}{Number of cores.}
}
\value{
A \code{gp3bayes_loo_weights} object.
}
\description{
Compute LOO Model-Averaging Weights
}
