% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-registry-dashboard.R
\name{create_diagnostic_dashboard_figures}
\alias{create_diagnostic_dashboard_figures}
\title{Create Diagnostic Dashboard Figures}
\usage{
create_diagnostic_dashboard_figures(x)
}
\arguments{
\item{x}{A diagnostic dashboard.}
}
\value{
A \code{gp3bayes_figure_set} for supported evidence components.
}
\description{
Create Diagnostic Dashboard Figures
}
