% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-registry-dashboard.R
\name{create_publication_registry}
\alias{create_publication_registry}
\title{Create a Publication Registry}
\usage{
create_publication_registry(label = NULL)
}
\arguments{
\item{label}{Optional registry label.}
}
\value{
A \code{gp3bayes_publication_registry}.
}
\description{
Create a Publication Registry
}
