% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-card.R
\name{create_reporting_checklist}
\alias{create_reporting_checklist}
\title{Create a Reporting Checklist}
\usage{
create_reporting_checklist(x)
}
\arguments{
\item{x}{A fitted \code{gp3bayes_fit} or \code{gp3bayes_model_card}.}
}
\value{
A data frame describing evidence that is available for reporting.
}
\description{
Create a Reporting Checklist
}
