% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-registry-dashboard.R
\name{diagnostic_dashboard_table}
\alias{diagnostic_dashboard_table}
\title{Diagnostic Dashboard Table}
\usage{
diagnostic_dashboard_table(x)
}
\arguments{
\item{x}{A diagnostic dashboard.}
}
\value{
Component availability and status.
}
\description{
Diagnostic Dashboard Table
}
