% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-support.R
\name{duration_pit_table}
\alias{duration_pit_table}
\title{Duration Probability-Integral-Transform Table}
\usage{
duration_pit_table(x)
}
\arguments{
\item{x}{A duration posterior predictive \code{gp3bayes_prediction}.}
}
\value{
Observation-level empirical posterior predictive PIT values.
}
\description{
Duration Probability-Integral-Transform Table
}
