% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{estimand_sensitivity_table}
\alias{estimand_sensitivity_table}
\title{Estimand Sensitivity Table}
\usage{
estimand_sensitivity_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_estimand_sensitivity}.}
}
\value{
Alternative-versus-reference estimand summaries.
}
\description{
Estimand Sensitivity Table
}
