% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-response-shape.R
\name{estimate_pupil_response_parameters}
\alias{estimate_pupil_response_parameters}
\title{Estimate nonlinear response-shape parameters}
\usage{
estimate_pupil_response_parameters(fit, probability = 0.95)
}
\arguments{
\item{fit}{A response-shape fit.}

\item{probability}{Central interval probability.}
}
\value{
A \code{gp3bayes_pupil_response_parameters} object.
}
\description{
Estimate nonlinear response-shape parameters
}
