% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-registry-dashboard.R
\name{evidence_inventory_table}
\alias{evidence_inventory_table}
\title{Evidence Inventory Table}
\usage{
evidence_inventory_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_evidence_inventory}.}
}
\value{
Inventory metadata.
}
\description{
Evidence Inventory Table
}
