% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfit-exploration.R
\name{extract_sampler_diagnostics}
\alias{extract_sampler_diagnostics}
\title{Extract NUTS Sampler Diagnostics}
\usage{
extract_sampler_diagnostics(fit)
}
\arguments{
\item{fit}{A fitted \code{gp3bayes_fit}.}
}
\value{
A data frame returned from \code{brms::nuts_params()}.
}
\description{
Extract NUTS Sampler Diagnostics
}
