% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-fit.R
\name{fit_advanced_pupil_model_cmdstanr}
\alias{fit_advanced_pupil_model_cmdstanr}
\title{Fit an advanced pupil model using cmdstanr}
\usage{
fit_advanced_pupil_model_cmdstanr(
  specification,
  chains = 4L,
  iter = 2000L,
  warmup = 1000L,
  cores = min(2L, chains),
  seed = 2026,
  adapt_delta = 0.95,
  max_treedepth = 12L,
  refresh = 0L
)
}
\arguments{
\item{specification}{An advanced pupil specification.}

\item{chains, iter, warmup, cores, seed, adapt_delta, max_treedepth, refresh}{Restricted MCMC controls.}
}
\description{
Fit an advanced pupil model using cmdstanr
}
