% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-optional-workflows.R
\name{interaction_prior_summary}
\alias{interaction_prior_summary}
\title{Summarise Separate Interaction Prior Metadata}
\usage{
interaction_prior_summary(specification)
}
\arguments{
\item{specification}{An advanced binary or duration specification.}
}
\value{
A one-row data frame.
}
\description{
Summarise Separate Interaction Prior Metadata
}
