% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfit-exploration.R
\name{loo_diagnostic_table}
\alias{loo_diagnostic_table}
\title{LOO Diagnostic Table}
\usage{
loo_diagnostic_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_psis_loo} or \code{loo} object.}
}
\value{
Observation-level Pareto-k diagnostics.
}
\description{
LOO Diagnostic Table
}
