% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfit-exploration.R
\name{loo_summary_table}
\alias{loo_summary_table}
\title{LOO Summary Table}
\usage{
loo_summary_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_psis_loo} or \code{loo} object.}
}
\value{
A tidy table of LOO estimates and standard errors.
}
\description{
LOO Summary Table
}
