% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{manifest_comparison_table}
\alias{manifest_comparison_table}
\title{Manifest-Comparison Table}
\usage{
manifest_comparison_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_manifest_comparison}.}
}
\value{
Manifest component comparisons.
}
\description{
Manifest-Comparison Table
}
