% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{missingness_audit_table}
\alias{missingness_audit_table}
\title{Missingness-Audit Table}
\usage{
missingness_audit_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_missingness_audit}.}
}
\value{
Column-level missingness diagnostics.
}
\description{
Missingness-Audit Table
}
