% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfit-exploration.R
\name{model_comparison_table}
\alias{model_comparison_table}
\title{LOO Model-Comparison Table}
\usage{
model_comparison_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_loo_comparison} or matrix returned by
\code{loo::loo_compare()}.}
}
\value{
A data frame retaining ELPD differences and their standard errors.
}
\description{
LOO Model-Comparison Table
}
