% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{model_evidence_table}
\alias{model_evidence_table}
\title{Model-Evidence Table}
\usage{
model_evidence_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_model_evidence}.}
}
\value{
The model-evidence inventory table.
}
\description{
Model-Evidence Table
}
