% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_advanced_pupil_simulation}
\alias{plot_advanced_pupil_simulation}
\title{Plot simulated advanced pupil trajectories}
\usage{
plot_advanced_pupil_simulation(x, observed = TRUE, ...)
}
\arguments{
\item{x}{An advanced pupil simulation.}

\item{observed}{If TRUE, plot observed condition means; otherwise plot stored
latent means.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot simulated advanced pupil trajectories
}
