% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_advanced_pupil_trajectory}
\alias{plot_advanced_pupil_trajectory}
\title{Plot an advanced posterior pupil trajectory}
\usage{
plot_advanced_pupil_trajectory(x, probability = 0.95, ...)
}
\arguments{
\item{x}{An advanced trajectory prediction.}

\item{probability}{Central interval probability.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot an advanced posterior pupil trajectory
}
