% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{plot_backend_environment_gg}
\alias{plot_backend_environment_gg}
\title{Plot Backend Environment Checks with ggplot2}
\usage{
plot_backend_environment_gg(x)
}
\arguments{
\item{x}{A \code{gp3bayes_backend_environment}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Backend Environment Checks with ggplot2
}
