% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{plot_backend_parity_gg}
\alias{plot_backend_parity_gg}
\title{Plot Backend Parity with ggplot2}
\usage{
plot_backend_parity_gg(x)
}
\arguments{
\item{x}{A \code{gp3bayes_backend_parity_audit}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Backend Parity with ggplot2
}
