% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_binary_calibration}
\alias{plot_binary_calibration}
\title{Binary Calibration Plot}
\usage{
plot_binary_calibration(x, bins = 10L)
}
\arguments{
\item{x}{A binary calibration table or binary expected prediction.}

\item{bins}{Number of bins if \code{x} is a prediction object.}
}
\value{
A \code{ggplot} object.
}
\description{
Binary Calibration Plot
}
