% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-distribution-atlas.R
\name{plot_binary_calibration_uncertainty}
\alias{plot_binary_calibration_uncertainty}
\title{Plot Binary Calibration Uncertainty}
\usage{
plot_binary_calibration_uncertainty(x)
}
\arguments{
\item{x}{A binary calibration-uncertainty object or table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Binary Calibration Uncertainty
}
