% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_binary_group_calibration}
\alias{plot_binary_group_calibration}
\title{Grouped Binary Calibration Plot}
\usage{
plot_binary_group_calibration(x, group = NULL)
}
\arguments{
\item{x}{Grouped calibration table or binary expected prediction.}

\item{group}{Grouping-column name when \code{x} is a prediction object.}
}
\value{
A \code{ggplot}.
}
\description{
Grouped Binary Calibration Plot
}
