% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_binary_precision_recall}
\alias{plot_binary_precision_recall}
\title{Binary Precision-Recall Plot}
\usage{
plot_binary_precision_recall(x, observed = NULL)
}
\arguments{
\item{x}{A precision-recall table, binary expected prediction, or numeric
probabilities.}

\item{observed}{Optional observed outcomes for numeric probabilities.}
}
\value{
A \code{ggplot}.
}
\description{
Binary Precision-Recall Plot
}
