% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_binary_roc}
\alias{plot_binary_roc}
\title{Binary ROC Plot}
\usage{
plot_binary_roc(x, observed = NULL)
}
\arguments{
\item{x}{A ROC table, binary expected prediction, or numeric probabilities.}

\item{observed}{Optional observed outcomes for numeric probabilities.}
}
\value{
A \code{ggplot}.
}
\description{
Binary ROC Plot
}
