% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_binocular_pupil_trajectory}
\alias{plot_binocular_pupil_trajectory}
\title{Plot binocular posterior trajectories}
\usage{
plot_binocular_pupil_trajectory(x, probability = x$probability, ...)
}
\arguments{
\item{x}{A binocular trajectory.}

\item{probability}{Central interval probability.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot binocular posterior trajectories
}
