% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{plot_design_support_gg}
\alias{plot_design_support_gg}
\title{Plot Design-Support Components with ggplot2}
\usage{
plot_design_support_gg(x)
}
\arguments{
\item{x}{A \code{gp3bayes_design_support_audit}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Design-Support Components with ggplot2
}
