% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-registry-dashboard.R
\name{plot_diagnostic_dashboard}
\alias{plot_diagnostic_dashboard}
\title{Plot Diagnostic Dashboard Availability}
\usage{
plot_diagnostic_dashboard(x)
}
\arguments{
\item{x}{A diagnostic dashboard.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Diagnostic Dashboard Availability
}
