% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_duration_pit}
\alias{plot_duration_pit}
\title{Duration PIT Plot}
\usage{
plot_duration_pit(x, bins = 10L)
}
\arguments{
\item{x}{A duration PIT table or duration posterior predictive object.}

\item{bins}{Number of histogram bins.}
}
\value{
A \code{ggplot} object.
}
\description{
Duration PIT Plot
}
