% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_duration_qq}
\alias{plot_duration_qq}
\title{Duration Predictive Q-Q Plot}
\usage{
plot_duration_qq(x)
}
\arguments{
\item{x}{A duration Q-Q table or duration posterior predictive object.}
}
\value{
A \code{ggplot}.
}
\description{
Duration Predictive Q-Q Plot
}
