% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_duration_quantile_calibration}
\alias{plot_duration_quantile_calibration}
\title{Duration Quantile-Calibration Plot}
\usage{
plot_duration_quantile_calibration(x)
}
\arguments{
\item{x}{A duration quantile-calibration table or duration predictive object.}
}
\value{
A \code{ggplot} object.
}
\description{
Duration Quantile-Calibration Plot
}
