% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_duration_tail}
\alias{plot_duration_tail}
\title{Duration Tail-Check Plot}
\usage{
plot_duration_tail(x)
}
\arguments{
\item{x}{A duration tail-check table.}
}
\value{
A \code{ggplot}.
}
\description{
Duration Tail-Check Plot
}
