% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_estimand_intervals}
\alias{plot_estimand_intervals}
\title{Estimand Interval Plot}
\usage{
plot_estimand_intervals(x, quantities = NULL, probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{x}{A \code{gp3bayes_estimand}.}

\item{quantities}{Optional estimand quantities.}

\item{probs}{Posterior interval probabilities.}
}
\value{
A \code{ggplot} object.
}
\description{
Estimand Interval Plot
}
