% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{plot_estimand_sensitivity_gg}
\alias{plot_estimand_sensitivity_gg}
\title{Plot Estimand Sensitivity with ggplot2}
\usage{
plot_estimand_sensitivity_gg(x)
}
\arguments{
\item{x}{An estimand-sensitivity object or its table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Estimand Sensitivity with ggplot2
}
