% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_exceedance_probability}
\alias{plot_exceedance_probability}
\title{Prediction Exceedance-Probability Plot}
\usage{
plot_exceedance_probability(x)
}
\arguments{
\item{x}{A table returned by \code{prediction_exceedance_probability()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Prediction Exceedance-Probability Plot
}
