% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{plot_group_deletion_sensitivity}
\alias{plot_group_deletion_sensitivity}
\title{Plot Group-Deletion Sensitivity}
\usage{
plot_group_deletion_sensitivity(x)
}
\arguments{
\item{x}{A group-deletion object or summary table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Group-Deletion Sensitivity
}
