% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical-effects-advanced.R
\name{plot_group_effect_rank_probability}
\alias{plot_group_effect_rank_probability}
\title{Plot Group-Effect Rank Probabilities}
\usage{
plot_group_effect_rank_probability(x)
}
\arguments{
\item{x}{A rank-probability table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Group-Effect Rank Probabilities
}
