% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_group_predictions}
\alias{plot_group_predictions}
\title{Group Prediction Plot}
\usage{
plot_group_predictions(x, group_column)
}
\arguments{
\item{x}{A group prediction summary.}

\item{group_column}{Name of the grouping column to use on the axis.}
}
\value{
A \code{ggplot}.
}
\description{
Group Prediction Plot
}
