% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_grouped_prediction_check}
\alias{plot_grouped_prediction_check}
\title{Grouped Posterior Predictive Plot}
\usage{
plot_grouped_prediction_check(x)
}
\arguments{
\item{x}{A \code{gp3bayes_group_prediction_check}.}
}
\value{
A \code{ggplot} object.
}
\description{
Grouped Posterior Predictive Plot
}
