% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-group-influence.R
\name{plot_loo_group_elpd}
\alias{plot_loo_group_elpd}
\title{Plot Grouped Pointwise ELPD Contribution}
\usage{
plot_loo_group_elpd(x)
}
\arguments{
\item{x}{A grouped LOO influence table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Grouped Pointwise ELPD Contribution
}
