% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-group-influence.R
\name{plot_loo_group_influence}
\alias{plot_loo_group_influence}
\title{Plot Grouped LOO Influence}
\usage{
plot_loo_group_influence(x)
}
\arguments{
\item{x}{A grouped LOO influence table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Grouped LOO Influence
}
