% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_loo_influence}
\alias{plot_loo_influence}
\title{LOO Influence Plot}
\usage{
plot_loo_influence(x)
}
\arguments{
\item{x}{A gp3bayes PSIS-LOO object, \code{loo} object, or LOO diagnostic table.}
}
\value{
A \code{ggplot} object.
}
\description{
LOO Influence Plot
}
