% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-influence-atlas.R
\name{plot_loo_influence_rank}
\alias{plot_loo_influence_rank}
\title{Plot Ranked LOO Influence}
\usage{
plot_loo_influence_rank(x)
}
\arguments{
\item{x}{A LOO object or pointwise table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Ranked LOO Influence
}
