% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-influence-atlas.R
\name{plot_loo_pointwise_elpd}
\alias{plot_loo_pointwise_elpd}
\title{Plot Pointwise LOO ELPD}
\usage{
plot_loo_pointwise_elpd(x)
}
\arguments{
\item{x}{A LOO object or pointwise table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Pointwise LOO ELPD
}
