% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{plot_manifest_comparison_gg}
\alias{plot_manifest_comparison_gg}
\title{Plot Manifest Differences with ggplot2}
\usage{
plot_manifest_comparison_gg(x)
}
\arguments{
\item{x}{A \code{gp3bayes_manifest_comparison}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Manifest Differences with ggplot2
}
