% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{plot_missingness_gg}
\alias{plot_missingness_gg}
\title{Plot Missingness Fractions with ggplot2}
\usage{
plot_missingness_gg(x)
}
\arguments{
\item{x}{A \code{gp3bayes_missingness_audit}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Missingness Fractions with ggplot2
}
