% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{LOO Model-Comparison Plot}
\usage{
plot_model_comparison(x)
}
\arguments{
\item{x}{A gp3bayes LOO comparison or comparison table.}
}
\value{
A \code{ggplot} object.
}
\description{
LOO Model-Comparison Plot
}
