% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{plot_model_evidence_gg}
\alias{plot_model_evidence_gg}
\title{Plot Model-Evidence Availability with ggplot2}
\usage{
plot_model_evidence_gg(x)
}
\arguments{
\item{x}{A \code{gp3bayes_model_evidence}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Model-Evidence Availability with ggplot2
}
