% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_model_weights}
\alias{plot_model_weights}
\title{LOO Model-Weight Plot}
\usage{
plot_model_weights(x)
}
\arguments{
\item{x}{A gp3bayes LOO weight object or weight table.}
}
\value{
A \code{ggplot} object.
}
\description{
LOO Model-Weight Plot
}
