% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_posterior_correlations}
\alias{plot_posterior_correlations}
\title{Posterior Correlation Plot}
\usage{
plot_posterior_correlations(
  x,
  variables = NULL,
  regex = NULL,
  method = c("pearson", "spearman")
)
}
\arguments{
\item{x}{A gp3bayes fit or posterior draws.}

\item{variables, regex}{Posterior variable selectors.}

\item{method}{Correlation method.}
}
\value{
A \code{ggplot} heatmap of posterior-draw correlations.
}
\description{
Posterior Correlation Plot
}
