% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_posterior_density}
\alias{plot_posterior_density}
\title{Posterior Density Plot}
\usage{
plot_posterior_density(x, variables = NULL, regex = NULL)
}
\arguments{
\item{x}{A gp3bayes fit or posterior draws.}

\item{variables, regex}{Posterior variable selectors.}
}
\value{
A \code{ggplot} object.
}
\description{
Posterior Density Plot
}
