% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_posterior_pairs}
\alias{plot_posterior_pairs}
\title{Posterior Pairs Plot}
\usage{
plot_posterior_pairs(fit, variables = NULL, regex = "^b_", max_variables = 8L)
}
\arguments{
\item{fit}{A fitted \code{gp3bayes_fit}.}

\item{variables}{Optional variables.}

\item{regex}{Optional variable regular expression.}

\item{max_variables}{Maximum number of variables displayed.}
}
\value{
A bayesplot pairs object.
}
\description{
Posterior Pairs Plot
}
