% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{plot_powerscale_sensitivity_gg}
\alias{plot_powerscale_sensitivity_gg}
\title{Plot Power-Scale Sensitivity with ggplot2}
\usage{
plot_powerscale_sensitivity_gg(x)
}
\arguments{
\item{x}{A power-scale sensitivity object or its table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Power-Scale Sensitivity with ggplot2
}
