% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_ppc_statistic}
\alias{plot_ppc_statistic}
\title{Posterior Predictive Statistic Plot}
\usage{
plot_ppc_statistic(x, bins = 30L)
}
\arguments{
\item{x}{A \code{gp3bayes_ppc_statistic}.}

\item{bins}{Histogram bins.}
}
\value{
A \code{ggplot}.
}
\description{
Posterior Predictive Statistic Plot
}
