% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-surfaces.R
\name{plot_prediction_contrast_profile}
\alias{plot_prediction_contrast_profile}
\title{Plot a Prediction Contrast Profile}
\usage{
plot_prediction_contrast_profile(x)
}
\arguments{
\item{x}{A prediction contrast profile.}
}
\value{
A \code{ggplot}.
}
\description{
Plot a Prediction Contrast Profile
}
