% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-surfaces.R
\name{plot_prediction_gradient}
\alias{plot_prediction_gradient}
\title{Plot Prediction-Profile Gradients}
\usage{
plot_prediction_gradient(x)
}
\arguments{
\item{x}{A prediction profile or gradient table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Prediction-Profile Gradients
}
