% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_prediction_interval_width}
\alias{plot_prediction_interval_width}
\title{Prediction Interval-Width Plot}
\usage{
plot_prediction_interval_width(x)
}
\arguments{
\item{x}{A prediction object or interval-width table.}
}
\value{
A \code{ggplot}.
}
\description{
Prediction Interval-Width Plot
}
