% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_prediction_intervals}
\alias{plot_prediction_intervals}
\title{Prediction-Interval Plot}
\usage{
plot_prediction_intervals(x, max_rows = 100L)
}
\arguments{
\item{x}{A \code{gp3bayes_prediction}.}

\item{max_rows}{Maximum rows displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Prediction-Interval Plot
}
