% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-surfaces.R
\name{plot_prediction_profile}
\alias{plot_prediction_profile}
\title{Plot a Prediction Profile}
\usage{
plot_prediction_profile(x)
}
\arguments{
\item{x}{A prediction profile.}
}
\value{
A \code{ggplot}.
}
\description{
Plot a Prediction Profile
}
