% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-predictive-graphics.R
\name{plot_prediction_rank_probabilities}
\alias{plot_prediction_rank_probabilities}
\title{Prediction Rank-Probability Plot}
\usage{
plot_prediction_rank_probabilities(x)
}
\arguments{
\item{x}{A ranking-probability table.}
}
\value{
A \code{ggplot}.
}
\description{
Prediction Rank-Probability Plot
}
