% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-distribution-atlas.R
\name{plot_prediction_score_uncertainty}
\alias{plot_prediction_score_uncertainty}
\title{Plot Prediction Score Uncertainty}
\usage{
plot_prediction_score_uncertainty(x)
}
\arguments{
\item{x}{A score-uncertainty object.}
}
\value{
A faceted \code{ggplot}.
}
\description{
Plot Prediction Score Uncertainty
}
