% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-surfaces.R
\name{plot_prediction_surface}
\alias{plot_prediction_surface}
\title{Plot a Prediction Surface}
\usage{
plot_prediction_surface(x)
}
\arguments{
\item{x}{A prediction surface.}
}
\value{
A \code{ggplot}.
}
\description{
Plot a Prediction Surface
}
