% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-surfaces.R
\name{plot_prediction_surface_uncertainty}
\alias{plot_prediction_surface_uncertainty}
\title{Plot Prediction-Surface Uncertainty}
\usage{
plot_prediction_surface_uncertainty(x)
}
\arguments{
\item{x}{A prediction surface.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Prediction-Surface Uncertainty
}
