% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-distribution-atlas.R
\name{plot_predictive_atlas_statistics}
\alias{plot_predictive_atlas_statistics}
\title{Plot Predictive Atlas Statistics}
\usage{
plot_predictive_atlas_statistics(x)
}
\arguments{
\item{x}{A predictive distribution atlas.}
}
\value{
A faceted \code{ggplot}.
}
\description{
Plot Predictive Atlas Statistics
}
