% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_predictive_coverage}
\alias{plot_predictive_coverage}
\title{Predictive-Coverage Plot}
\usage{
plot_predictive_coverage(x)
}
\arguments{
\item{x}{Predictive coverage table or posterior predictive object.}
}
\value{
A \code{ggplot} object.
}
\description{
Predictive-Coverage Plot
}
