% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-distribution-atlas.R
\name{plot_predictive_quantile_envelope}
\alias{plot_predictive_quantile_envelope}
\title{Plot Posterior-Predictive Quantile Envelope}
\usage{
plot_predictive_quantile_envelope(x)
}
\arguments{
\item{x}{A quantile-envelope table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Posterior-Predictive Quantile Envelope
}
