% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_predictive_residuals}
\alias{plot_predictive_residuals}
\title{Predictive-Residual Plot}
\usage{
plot_predictive_residuals(x)
}
\arguments{
\item{x}{Residual table returned by \code{predictive_residuals()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Predictive-Residual Plot
}
