% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-posterior-bridge.R
\name{plot_prior_posterior_contraction}
\alias{plot_prior_posterior_contraction}
\title{Plot Prior-to-Posterior Contraction}
\usage{
plot_prior_posterior_contraction(x)
}
\arguments{
\item{x}{A prior-posterior bridge.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Prior-to-Posterior Contraction
}
