% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-posterior-bridge.R
\name{plot_prior_posterior_intervals}
\alias{plot_prior_posterior_intervals}
\title{Plot Prior and Posterior Intervals}
\usage{
plot_prior_posterior_intervals(x)
}
\arguments{
\item{x}{A prior-posterior bridge.}
}
\value{
A faceted \code{ggplot}.
}
\description{
Plot Prior and Posterior Intervals
}
