% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{plot_prior_sensitivity_scenarios}
\alias{plot_prior_sensitivity_scenarios}
\title{Plot Prior-Sensitivity Scenario Maxima}
\usage{
plot_prior_sensitivity_scenarios(x)
}
\arguments{
\item{x}{A prior-sensitivity object or scenario table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Prior-Sensitivity Scenario Maxima
}
