% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_autocorrelation_comparison}
\alias{plot_pupil_autocorrelation_comparison}
\title{Plot residual autocorrelation comparison}
\usage{
plot_pupil_autocorrelation_comparison(x, absolute = TRUE, ...)
}
\arguments{
\item{x}{An autocorrelation comparison object.}

\item{absolute}{Plot median absolute ACF if TRUE.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot residual autocorrelation comparison
}
