% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_dynamic_contrast}
\alias{plot_pupil_dynamic_contrast}
\title{Plot a dynamic pupil condition contrast}
\usage{
plot_pupil_dynamic_contrast(x, ...)
}
\arguments{
\item{x}{A dynamic contrast.}

\item{...}{Additional base graphics arguments.}
}
\description{
Plot a dynamic pupil condition contrast
}
