% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-graphics.R
\name{plot_pupil_estimand}
\alias{plot_pupil_estimand}
\title{Plot declared-window pupil estimands}
\usage{
plot_pupil_estimand(x)
}
\arguments{
\item{x}{Pupil estimand object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot declared-window pupil estimands
}
\examples{
# See vignette("bayesian-dynamic-pupillometry", package = "gp3bayes") for a complete workflow.
}
