% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_gp_hyperparameters}
\alias{plot_pupil_gp_hyperparameters}
\title{Plot Gaussian-process hyperparameters}
\usage{
plot_pupil_gp_hyperparameters(x, ...)
}
\arguments{
\item{x}{A GP-hyperparameter object.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot Gaussian-process hyperparameters
}
