% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_identifiability_audit}
\alias{plot_pupil_identifiability_audit}
\title{Plot advanced identifiability/design-support audit}
\usage{
plot_pupil_identifiability_audit(x, ...)
}
\arguments{
\item{x}{An identifiability audit.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot advanced identifiability/design-support audit
}
