% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_lfo}
\alias{plot_pupil_lfo}
\title{Plot leave-future-out scores}
\usage{
plot_pupil_lfo(x, ...)
}
\arguments{
\item{x}{An executed LFO validation or LFO comparison.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot leave-future-out scores
}
