% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_measurement_uncertainty}
\alias{plot_pupil_measurement_uncertainty}
\title{Plot measurement-uncertainty audit}
\usage{
plot_pupil_measurement_uncertainty(x, ...)
}
\arguments{
\item{x}{A measurement audit.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot measurement-uncertainty audit
}
