% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_missingness}
\alias{plot_pupil_missingness}
\title{Plot missing-response fraction over time}
\usage{
plot_pupil_missingness(x, ...)
}
\arguments{
\item{x}{A missingness audit.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot missing-response fraction over time
}
