% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_model_comparison}
\alias{plot_pupil_model_comparison}
\title{Plot predictive model comparison}
\usage{
plot_pupil_model_comparison(x, ...)
}
\arguments{
\item{x}{A model-comparison object.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot predictive model comparison
}
