% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_model_complexity}
\alias{plot_pupil_model_complexity}
\title{Plot advanced model computational complexity}
\usage{
plot_pupil_model_complexity(x, ...)
}
\arguments{
\item{x}{A complexity audit or advanced specification.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot advanced model computational complexity
}
