% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_predictive_calibration}
\alias{plot_pupil_predictive_calibration}
\title{Plot posterior predictive calibration}
\usage{
plot_pupil_predictive_calibration(x, ...)
}
\arguments{
\item{x}{A predictive score/calibration object.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot posterior predictive calibration
}
