% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_residual_scale}
\alias{plot_pupil_residual_scale}
\title{Plot posterior residual scale over time}
\usage{
plot_pupil_residual_scale(x, ...)
}
\arguments{
\item{x}{A residual-scale estimand.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot posterior residual scale over time
}
