% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_residual_spectrum}
\alias{plot_pupil_residual_spectrum}
\title{Plot residual spectrum}
\usage{
plot_pupil_residual_spectrum(x, ...)
}
\arguments{
\item{x}{A residual-spectrum object.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot residual spectrum
}
