% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_response_parameters}
\alias{plot_pupil_response_parameters}
\title{Plot experimental nonlinear response parameters}
\usage{
plot_pupil_response_parameters(x, ...)
}
\arguments{
\item{x}{A response-parameter object.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot experimental nonlinear response parameters
}
