% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_temporal_dependence}
\alias{plot_pupil_temporal_dependence}
\title{Plot empirical temporal-dependence audit}
\usage{
plot_pupil_temporal_dependence(x, ...)
}
\arguments{
\item{x}{A temporal-dependence audit.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot empirical temporal-dependence audit
}
