% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-graphics.R
\name{plot_pupil_trajectory_derivative}
\alias{plot_pupil_trajectory_derivative}
\title{Plot posterior pupil trajectory derivative}
\usage{
plot_pupil_trajectory_derivative(x, probability = x$probability, ...)
}
\arguments{
\item{x}{A trajectory-derivative object.}

\item{probability}{Central posterior interval probability.}

\item{...}{Additional graphical arguments.}
}
\description{
Plot posterior pupil trajectory derivative
}
