% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-graphics.R
\name{plot_pupil_validation}
\alias{plot_pupil_validation}
\title{Plot pupil validation evidence}
\usage{
plot_pupil_validation(x)
}
\arguments{
\item{x}{Pupil validation object.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot pupil validation evidence
}
\examples{
# See vignette("bayesian-dynamic-pupillometry", package = "gp3bayes") for a complete workflow.
}
