% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical-effects-advanced.R
\name{plot_random_intercept_variance_partition}
\alias{plot_random_intercept_variance_partition}
\title{Plot Random-Intercept Variance Partition}
\usage{
plot_random_intercept_variance_partition(x)
}
\arguments{
\item{x}{A variance-partition object or table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Random-Intercept Variance Partition
}
