% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{plot_recovery_bias}
\alias{plot_recovery_bias}
\title{Plot Recovery Standardized Bias}
\usage{
plot_recovery_bias(x)
}
\arguments{
\item{x}{A recovery object or recovery parameter table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Recovery Standardized Bias
}
