% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{plot_recovery_coverage}
\alias{plot_recovery_coverage}
\title{Plot Recovery Coverage}
\usage{
plot_recovery_coverage(x)
}
\arguments{
\item{x}{A recovery object or recovery parameter table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Recovery Coverage
}
