% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{plot_recovery_rmse}
\alias{plot_recovery_rmse}
\title{Plot Recovery RMSE}
\usage{
plot_recovery_rmse(x)
}
\arguments{
\item{x}{A recovery object or recovery parameter table.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Recovery RMSE
}
