% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-card.R
\name{plot_reporting_checklist}
\alias{plot_reporting_checklist}
\title{Plot a Reporting Checklist}
\usage{
plot_reporting_checklist(x)
}
\arguments{
\item{x}{A reporting checklist, fit, or model card.}
}
\value{
A \code{ggplot}.
}
\description{
Plot a Reporting Checklist
}
