% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{plot_schema_comparison_gg}
\alias{plot_schema_comparison_gg}
\title{Plot Schema Differences with ggplot2}
\usage{
plot_schema_comparison_gg(x)
}
\arguments{
\item{x}{A \code{gp3bayes_schema_comparison}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot Schema Differences with ggplot2
}
