% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{plot_sensitivity_suite_gg}
\alias{plot_sensitivity_suite_gg}
\title{Plot a Sensitivity Suite with ggplot2}
\usage{
plot_sensitivity_suite_gg(x)
}
\arguments{
\item{x}{A \code{gp3bayes_sensitivity_suite}.}
}
\value{
A \code{ggplot}.
}
\description{
Plot a Sensitivity Suite with ggplot2
}
