% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_uncertainty_decomposition}
\alias{plot_uncertainty_decomposition}
\title{Prediction-Uncertainty Plot}
\usage{
plot_uncertainty_decomposition(x, max_rows = 100L)
}
\arguments{
\item{x}{A \code{gp3bayes_prediction_uncertainty}.}

\item{max_rows}{Maximum observations displayed.}
}
\value{
A \code{ggplot} object.
}
\description{
Prediction-Uncertainty Plot
}
