% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{plot_variance_components}
\alias{plot_variance_components}
\title{Variance-Component Plot}
\usage{
plot_variance_components(x)
}
\arguments{
\item{x}{A variance-component table or fitted gp3bayes object.}
}
\value{
A \code{ggplot} object.
}
\description{
Variance-Component Plot
}
