% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postfit-exploration.R
\name{posterior_correlation_table}
\alias{posterior_correlation_table}
\title{Posterior Correlation Table}
\usage{
posterior_correlation_table(
  x,
  variables = NULL,
  regex = NULL,
  method = c("pearson", "spearman")
)
}
\arguments{
\item{x}{A gp3bayes fit or posterior draws accepted by
\code{posterior_interval_table()}.}

\item{variables, regex}{Posterior variable selectors.}

\item{method}{Correlation method.}
}
\value{
A long data frame of unique posterior-draw correlations.
}
\description{
Posterior Correlation Table
}
\examples{
x <- cbind(a = rnorm(100), b = rnorm(100), c = rnorm(100))
posterior_correlation_table(x)
}
