% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery-sensitivity-publication.R
\name{powerscale_sensitivity_table}
\alias{powerscale_sensitivity_table}
\title{Power-Scale Sensitivity Table}
\usage{
powerscale_sensitivity_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_powerscale_sensitivity}.}
}
\value{
The tabular representation provided by \code{priorsense}.
}
\description{
Power-Scale Sensitivity Table
}
