% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-diagnostics-advanced.R
\name{ppc_statistic_table}
\alias{ppc_statistic_table}
\title{Posterior Predictive Statistic Table}
\usage{
ppc_statistic_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_ppc_statistic}.}
}
\value{
A one-row data frame.
}
\description{
Posterior Predictive Statistic Table
}
