% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-surfaces.R
\name{prediction_contrast_profile_table}
\alias{prediction_contrast_profile_table}
\title{Prediction Contrast Profile Table}
\usage{
prediction_contrast_profile_table(x)
}
\arguments{
\item{x}{A prediction contrast profile.}
}
\value{
Contrast summaries.
}
\description{
Prediction Contrast Profile Table
}
