% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-diagnostics-advanced.R
\name{prediction_interval_width}
\alias{prediction_interval_width}
\title{Prediction Interval Width Table}
\usage{
prediction_interval_width(x)
}
\arguments{
\item{x}{A \code{gp3bayes_prediction}.}
}
\value{
A table of posterior interval width by prediction row.
}
\description{
Prediction Interval Width Table
}
