% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-surfaces.R
\name{prediction_profile_table}
\alias{prediction_profile_table}
\title{Prediction Profile Table}
\usage{
prediction_profile_table(x)
}
\arguments{
\item{x}{A prediction profile.}
}
\value{
Profile summaries.
}
\description{
Prediction Profile Table
}
