% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-distribution-atlas.R
\name{prediction_score_uncertainty_table}
\alias{prediction_score_uncertainty_table}
\title{Prediction Score-Uncertainty Table}
\usage{
prediction_score_uncertainty_table(x)
}
\arguments{
\item{x}{A prediction-score uncertainty object.}
}
\value{
Metric summaries.
}
\description{
Prediction Score-Uncertainty Table
}
