% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-support.R
\name{prediction_support_table}
\alias{prediction_support_table}
\title{Prediction-Support Table}
\usage{
prediction_support_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_prediction_support} object.}
}
\value{
The underlying support audit table.
}
\description{
Prediction-Support Table
}
