% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-surfaces.R
\name{prediction_surface_table}
\alias{prediction_surface_table}
\title{Prediction Surface Table}
\usage{
prediction_surface_table(x)
}
\arguments{
\item{x}{A prediction surface.}
}
\value{
Surface summaries.
}
\description{
Prediction Surface Table
}
