% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-support.R
\name{prediction_table}
\alias{prediction_table}
\title{Prediction Summary Table}
\usage{
prediction_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_prediction}.}
}
\value{
The observation-level posterior prediction summary.
}
\description{
Prediction Summary Table
}
