% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-support.R
\name{predictive_coverage_table}
\alias{predictive_coverage_table}
\title{Posterior Predictive Coverage Table}
\usage{
predictive_coverage_table(x, levels = c(0.5, 0.8, 0.9, 0.95))
}
\arguments{
\item{x}{A posterior predictive \code{gp3bayes_prediction}.}

\item{levels}{Central predictive interval levels.}
}
\value{
A table of empirical coverage and mean interval width.
}
\description{
Posterior Predictive Coverage Table
}
