% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-specification.R
\name{print.gp3bayes_prior_specification}
\alias{print.gp3bayes_prior_specification}
\title{Print a gp3bayes Prior Specification}
\usage{
\method{print}{gp3bayes_prior_specification}(x, ...)
}
\arguments{
\item{x}{A \code{gp3bayes_prior_specification} object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a gp3bayes Prior Specification
}
\keyword{internal}
