% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-posterior-bridge.R
\name{prior_posterior_distance_table}
\alias{prior_posterior_distance_table}
\title{Prior-versus-Posterior Distance Table}
\usage{
prior_posterior_distance_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_prior_posterior_bridge}.}
}
\value{
Marginal empirical distribution-distance summaries.
}
\description{
Prior-versus-Posterior Distance Table
}
