% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-registry-dashboard.R
\name{publication_registry_table}
\alias{publication_registry_table}
\title{Publication Registry Table}
\usage{
publication_registry_table(x)
}
\arguments{
\item{x}{A publication registry.}
}
\value{
Entry metadata.
}
\description{
Publication Registry Table
}
