% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-specification.R
\name{pupil_advanced_specification_table}
\alias{pupil_advanced_specification_table}
\title{Tabulate an advanced pupil model specification}
\usage{
pupil_advanced_specification_table(x)
}
\arguments{
\item{x}{An advanced specification.}
}
\value{
A one-row data frame.
}
\description{
Tabulate an advanced pupil model specification
}
