% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-binocular.R
\name{pupil_binocular_correlation}
\alias{pupil_binocular_correlation}
\title{Extract posterior residual binocular correlation}
\usage{
pupil_binocular_correlation(fit, probability = 0.95)
}
\arguments{
\item{fit}{A binocular fit with residual correlation enabled.}

\item{probability}{Central interval probability.}
}
\value{
A data frame.
}
\description{
Extract posterior residual binocular correlation
}
