% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-binocular.R
\name{pupil_binocular_difference}
\alias{pupil_binocular_difference}
\title{Estimate posterior right-minus-left binocular differences}
\usage{
pupil_binocular_difference(x)
}
\arguments{
\item{x}{A binocular trajectory object.}
}
\value{
A data frame.
}
\description{
Estimate posterior right-minus-left binocular differences
}
