% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-functional-estimands.R
\name{pupil_dynamic_contrast_table}
\alias{pupil_dynamic_contrast_table}
\title{Tabulate a dynamic pupil contrast}
\usage{
pupil_dynamic_contrast_table(x)
}
\arguments{
\item{x}{A dynamic-contrast object.}
}
\value{
A data frame.
}
\description{
Tabulate a dynamic pupil contrast
}
