% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-estimands.R
\name{pupil_gp_table}
\alias{pupil_gp_table}
\title{Tabulate GP hyperparameters}
\usage{
pupil_gp_table(x)
}
\arguments{
\item{x}{A GP-hyperparameter object.}
}
\value{
A data frame.
}
\description{
Tabulate GP hyperparameters
}
