% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-estimands.R
\name{pupil_missingness_table}
\alias{pupil_missingness_table}
\title{Tabulate missingness audit results}
\usage{
pupil_missingness_table(x)
}
\arguments{
\item{x}{A missingness audit.}
}
\value{
A data frame.
}
\description{
Tabulate missingness audit results
}
