% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-model-comparison.R
\name{pupil_model_comparison_table}
\alias{pupil_model_comparison_table}
\title{Tabulate model comparison}
\usage{
pupil_model_comparison_table(x)
}
\arguments{
\item{x}{A model-comparison object.}
}
\value{
A data frame.
}
\description{
Tabulate model comparison
}
