% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-advanced-estimands.R
\name{pupil_residual_scale_table}
\alias{pupil_residual_scale_table}
\title{Tabulate posterior residual scale}
\usage{
pupil_residual_scale_table(x)
}
\arguments{
\item{x}{A residual-scale estimand.}
}
\value{
A data frame.
}
\description{
Tabulate posterior residual scale
}
