% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-response-shape.R
\name{pupil_response_parameter_table}
\alias{pupil_response_parameter_table}
\title{Tabulate nonlinear response parameters}
\usage{
pupil_response_parameter_table(x)
}
\arguments{
\item{x}{A response-parameter object.}
}
\description{
Tabulate nonlinear response parameters
}
