% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-specification.R
\name{pupil_specification_table}
\alias{pupil_specification_table}
\title{Convert a pupil specification to a publication-ready table}
\usage{
pupil_specification_table(x)
}
\arguments{
\item{x}{A pupil model specification.}
}
\value{
A data frame.
}
\description{
Convert a pupil specification to a publication-ready table
}
\examples{
# See vignette("bayesian-dynamic-pupillometry", package = "gp3bayes") for a complete workflow.
}
