% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-validation.R
\name{pupil_validation_table}
\alias{pupil_validation_table}
\title{Extract a pupil validation table}
\usage{
pupil_validation_table(x)
}
\arguments{
\item{x}{A pupil validation object.}
}
\value{
A data frame.
}
\description{
Extract a pupil validation table
}
\examples{
# See vignette("bayesian-dynamic-pupillometry", package = "gp3bayes") for a complete workflow.
}
