% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-graphics-gg.R
\name{schema_comparison_table}
\alias{schema_comparison_table}
\title{Schema-Comparison Table}
\usage{
schema_comparison_table(x)
}
\arguments{
\item{x}{A \code{gp3bayes_schema_comparison}.}
}
\value{
Structural schema comparisons.
}
\description{
Schema-Comparison Table
}
