% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-optional-workflows.R
\name{summarise_sbc_result}
\alias{summarise_sbc_result}
\title{Summarise an SBC Result}
\usage{
summarise_sbc_result(x)
}
\arguments{
\item{x}{A \code{gp3bayes_sbc_result}.}
}
\value{
A conservative summary list.
}
\description{
Summarise an SBC Result
}
