% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-graphics.R
\name{theme_gp3bayes}
\alias{theme_gp3bayes}
\title{gp3bayes Publication Theme}
\usage{
theme_gp3bayes(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A \code{ggplot2} theme.
}
\description{
gp3bayes Publication Theme
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) theme_gp3bayes()
}
