% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupil-binocular.R
\name{translate_binocular_pupil_model_to_brms}
\alias{translate_binocular_pupil_model_to_brms}
\title{Translate a binocular specification to a brms multivariate formula}
\usage{
translate_binocular_pupil_model_to_brms(specification)
}
\arguments{
\item{specification}{A binocular specification.}
}
\value{
A \code{gp3bayes_binocular_brms_specification} object.
}
\description{
Translate a binocular specification to a brms multivariate formula
}
