% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced-optional-workflows.R
\name{translate_duration_model_with_interaction_prior}
\alias{translate_duration_model_with_interaction_prior}
\title{Translate an Advanced Duration Specification to brms}
\usage{
translate_duration_model_with_interaction_prior(specification)
}
\arguments{
\item{specification}{An advanced duration specification.}
}
\value{
A validated restricted brms translation.
}
\description{
Translate an Advanced Duration Specification to brms
}
