% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication-registry-dashboard.R
\name{validate_publication_registry}
\alias{validate_publication_registry}
\title{Validate a Publication Registry}
\usage{
validate_publication_registry(x)
}
\arguments{
\item{x}{A publication registry.}
}
\value{
A validation object.
}
\description{
Validate a Publication Registry
}
